requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pypdf>=3.17.0
python-docx>=1.1.0
pgvector>=0.2.0
django-cors-headers>=4.3.0
